Refactor functions to clarify that they are specific functions for org/project id logic
